Updated helpers which keep trying to run until completion

+++ bin/samba_smart_tmux.bash
@@ -17,9 +17,25 @@
 # Better idea, Let this be better handled elsewhere
 # reservation_name=jjc29_119
 # reservation_name="";
+limit=$1;
+confirm=$2;
+max_try=$3;
+
+if [ -z "$limit" ];then
+    # to make code easier, limit will be 100 by default, effectively no limit
+    limit=100;
+fi;
+if [ -z "$confirm" ];then
+    confirm=1;
+fi;
+if [ -z "$max_try" ]; then
+    max_try=10;
+fi;
+
 if [ ! -d samba_logs ];then
     mkdir samba_logs;fi;
-for shf in *.headfile; 
+
+for shf in $(ls -tr *.headfile); 
 do 
     # for the test sets the vbmsuffix was both in the file AND in the filename.
     #vbmsuffix=$(echo ${hf%.*}|cut -d '_' -f2);
@@ -31,6 +47,11 @@ do
     # result pattern
     rp="$BIGGUS_DISKUS/VBM*$vbmsuffix-results/*headfile";
     if [ $(tmux list-session 2> /dev/null |grep -c $vbmsuffix: 2>/dev/null) -le 0 ];then
+	# check limit; skip if nee dbe
+	if [ $limit -le 0 ];then
+	    #echo "Limit reached, skipping $vbmsuffix";
+	    continue;
+	fi;
 	# headfile path or empty when the file doesnt exist.
 	# Now that I look at things, might be able to do our ls check, 
 	# and capture status with $?, and test status
@@ -40,14 +61,28 @@ do
 	# If we didn't find a file using above ls command 
 	# OR, the one found is Older, launch.
 	if [ -z "$hfp" -o \( "$hfp" -ot $shf \) ]; then
-	    echo "Launching in 2 seconds ... ";
+	    st_count=$(grep -ci 'start work' samba_logs/$vbmsuffix.log 2> /dev/null || echo 0);
+	    if [ $st_count -gt $max_try ];then
+		if [ ! -d max_fail ];then mkdir max_fail; fi;
+		mv $shf max_fail/$(basename $shf); echo "too many failures for $vbmsuffix"; 
+		continue;
+	    fi;
 	    echo tmux new-session -d -s $vbmsuffix -- "\" source ~/.bashrc && SAMBA_startup $PWD/$shf 2>&1 | tee -a samba_logs/$vbmsuffix.log\"";
-	    sleep 2; tmux new-session -d -s $vbmsuffix -- " source ~/.bashrc && SAMBA_startup $PWD/$shf 2>&1 | tee -a samba_logs/$vbmsuffix.log";
-	    
-	else 
+	    if [ $confirm -eq 1 ];then
+		read -p "Start ${vbmsuffix}: $(basename $shf)(yN)" start_now
+	    else
+		start_now=y;
+	    fi;
+	    if [ "$start_now" == y -o "$start_now" == Y ];then 
+		echo "Launching in 2 seconds ... ";
+		sleep 2; tmux new-session -d -s $vbmsuffix -- " source ~/.bashrc && SAMBA_startup $PWD/$shf 2>&1 | tee -a samba_logs/$vbmsuffix.log";
+		let limit=$limit-1;
+	    fi;
+	else
 	    echo "Dat:$vbmsuffix complete $hfp";
 	fi;
     else
+	let limit=$limit-1;
 	echo "Dat:$vbmsuffix in progress";
     fi;
 done
